MBS-13268: Stop cleaning up CDBaby links
CD Baby is gone anyway, so added URLs nowadays will be from old data.
As mentioned in our forum
(https://community.metabrainz.org/t/cd-baby-store-closes-on-31-march/467757/12)
we often could benefit from not cleaning those since the cleanup
can make it harder to find archived versions of the data,
and the "cleaned" URL is also broken by now anyway.

This removes cleanup/validation, but keeps the autoselection
of the cdbaby reltype for artists.