MBS-13072: Notify on subsecond recording edit lengths
If an edit such as edit #99374885 only makes subsecond changes
to a recording's length, nothing at all is shown in the edit,
since we diff on seconds and as such the diff is empty.
That means it's entirely unclear what the edit even tries to do.

This shows an informative message equivalent to the one
in edit/details/SetTrackLengths, since we seem to consider
subsecond changes not worth displaying.