MBS-13244: Add standalone recording filter to artist index
When we fall back to showing recordings for an artist overview,
we still use the release group filter, which unsurprisingly is useless.

This makes it so that if the artist has no RGs, the filter loaded
becomes the recording one, which actually filters
the returned list of standalone recordings as well.