MBS-13145: Also show collection tab to logged out viewers
We currently require login to see any data about collections
for a specific entity, even when the entity is in a public collection
that can otherwise be seen when logged out.
I don't think this is really intentional, but more of a side effect
of the way the collections page / sidebar got implemented.
I asked Aerozol and he also suggested that this data
should be shown to everyone, since it's interesting to showcase.

This shows only the "in X collections" link on the sidebar
for logged out users (all others seem irrelevant).
It also allows them to see the /collections page for the entity,
showing only public collections.
We still want to pass show_private_only for _all_non_visible_collections
when there's no user, since otherwise it will not restrict it
to non-visible collections, but we don't have an editor ID to pass;
I chose to just pass 'none' as a string, then make it skip
the ID-related conditions if that is passed.