MBS-8952: Can rename recordings via autocomplete
I was able to reproduce this using the steps in
https://tickets.metabrainz.org/browse/MBS-13021.

The primary issue is that recordings from the search server don't include a
database row ID, so the `if (currentSelection.id)` check modified here didn't
clear the selection when modifying text. The `else` branch updated the
recording name with the entered text and notified all subscribers of the
autocomplete observable.

Checking for `currentSelection.gid` there fixes the bug. I also changed the
edit generation code to revert any recording changes if the necessary settings
to update them aren't enabled.