MBS-13279: Don't consider the empty string an email address
Previously has_email_address was defined as a predicate for whether or
not the email address was set, even if set as an empty string
(which is the case for a couple thousand accounts).

This patch removes the predicate and just defines has_email_address
as a subroutine that checks whether email is non-empty.

For reference, see the documentation for Moose predicates at
https://metacpan.org/dist/Moose/view/lib/Moose/Manual/Attributes.pod#Predicate-and-clearer-methods