MBS-13155: Don't consider all unselected labels dupes (#3046)
releaseLabelKey was automatically defaulting to '' as the label key
if no ID exists, meaning two labels such as "A" and "B" would get
seen as a dupe until an actual label with an ID was selected.
This has confused users in the past, since as far as they could tell
the labels were different.
This just makes it so that a label name is also taken into account
when calculating the key. I chose to add a prefix to it to avoid
the very unlikely but possible situation where a "label name"
that matches the other label's row ID would trigger a "duplicate" error.