Remove all uses of the smartmatch operator
The smartmatch operator (`~~`) is deprecated as of Perl 5.38.

https://perldoc.perl.org/perl5380delta#Switch-and-Smart-Match-operator

Although it had many use cases, it appears that we only used it for checking
that a value exists in an array.  I added two utility functions that help with
this: `contains_number` and `contains_string`.  I made these handle cases where
the array ref is `undef`, because the smartmatch operator also works in this
case, and I didn't verify that every input is defined.

In other cases, where the container array is a static list of strings, I
changed the code to perform a simple regex match.