MBS-13320: Prompt before unloading modified forms
Install a beforeunload event listener on the artist, event,
label, place, recording, release group, series, and work
edit pages that prompts the user if any of the form's inputs
have been changed. This is intended to prevent data loss
if the page is accidentally closed before the form has been
submitted.

One limitation of the approach used here is that it doesn't
detect relationship changes.

Note that the release editor already displays a prompt if
the page is unloaded with unsubmitted changes.