Normalize "pending edits" to "open edits" in UI
"pending edits" are, in the way we are using the term, exactly the same
as "open edits". As such, this normalizes the user-facing strings
so that we always use "open edits" when we mean that.

We could arguably use "pending edits" for edits that still
have not been entered, like the ones that will be created by an open
release editor or relationship editor session.
That would seem like a much better use of the term, but sadly
might require a huge refactoring
since we use "edits_pending" in our schema for open edits.