Regenerate cpanfile.snapshot
This is following the base-image upgrade to jammy.

Note: I've copied the existing cpanfile.snapshot to cpanfile.perl5.30.snapshot,
so that docker/Dockerfile.perl5.30.tests can continue to use it. The new
snapshot was generated from the jammy image containing Perl 5.34, and doesn't
seem to work in the old image.