Explicitly list the .gitignore'd Dockerfiles
We only intend to ignore Dockerfiles generated from .m4 templates.
Dockerfile.tests is not, though it matches the pattern changed here. It seems
an exception was made for it (since it's tracked by git), although I still have
trouble searching for Dockerfile.tests in tools that are aware of .gitignore,
like VS Code.