MBS-13347: Relationship Type edit search times out
The use of `LIMIT 500` inside an edit search sub-query using the "Relationship
Type" predicate causes PostgreSQL to ignore `edit_data_idx_link_type` and make
use of a much less efficient query plan.

This commit adds a flag to `EditSearch::Predicate` which can be used to disable
the addition of `$LIMIT_FOR_EDIT_LISTING` if the given predicate is present.
It's currently set for edit searches involving a condition on the type of the
relationship being edited (i.e., the "Relationship Type" predicate mentioned
above).