MBS-13348: Fix edit_data_idx_link_type (#3073)
The query in the edit search that is supposed to make use of this index was
changed in b21e5c579d0201325f5fbe385a51b663ae6caa69, but the actual index was
never updated to match (so it could not be used).

Note: This was already fixed in production. This patch provides an upgrade
script to run during the next schema change.