Allow connecting to any database with SYSTEM credentials
Sometimes it's necessary to connect to the database specified by MAINTENANCE or
READWRITE using SYSTEM credentials (because certain objects owned by a
superuser need to be updated). This is now possible by connecting to
SYSTEM_MAINTENANCE, SYSTEM_READWRITE, or any variant of SYSTEM_FOO, where FOO
is the name of a database defined in lib/DBDefs.pm.