MBS-13326: Add documentation bubble for the artist sort name field
This is a non-obvious field. Recently we've seen people reading it
as "short name", but even without that, it's often filled wrong.

There seems to be entirely no downsides to adding a bubble
with a short explanation and a link to the guidelines,
so I'm doing that and hoping it's enough to clarify the usage.