Add a read-only database connector
This adds an `ro_connector` to the server context which is intended to be used
to distribute read-only queries to our standby in production.

It must be enabled via `USE_RO_DATABASE_CONNECTOR`; more information can be
found in lib/DBDefs/Default.pm.

The `prefer_ro_*` methods were added to handle cases where use of the
`ro_connector` depends on whether we're in an existing writable transaction.