MBS-13273: Update track credits with matching prefixes (#3082)
When release artist credits are updated, also update track
credits that start with the old release credits (instead of
only ones that exactly match the release credits).

For example, if the release credits are changed from "A" to
"B", a track that was credited to "A feat. C" will be
updated to "B feat. C".