MBS-13055: Search instrument attributes by alias
This commit only affects searching for instrument attributes in the
relationship editor.

It makes use of the instrument aliases loaded by
bcec41b96ccbd4993d51864a54f22444fe750062 in searchItems.js.

We now attempt to match against either the instrument name or any alias name in
`weightEntry`. In fact, the aliases are weighted the same as the instrument
name now. This may be confusing, because you can't actually see what your
search term is matching against, but neither could you with the (external)
indexed search AFAIK.