MBS-13337: Use read-only connector for `Data::Role::QueryToList`
The main implication of using the standby instead of the primary here is
replication lag, which seems to hover between 50 and 400 ms at the time of
writing. However, note that `ro_connector` is not used if the user is
logged-in; so there should be no case where just-applied edits will not be
visible to them.