MBS-13368: Properly compare series relationships if series is target
We were only performing special series checks in compareRelationships
in cases where the series entity is the source. This means entering
relationships to series from other entities
was entirely skipping this mechanism.

This commit makes it so that the same mechanism is used if both targets
are the same series, and the other conditions apply.