Drop hard-coded IP address for incrental JSON dump
The IP address of the SSH server used to sync incremental JSON dumps
could not be changed without changing it in the current container, and
ultimately in the source code/image.

This patch instead uses the hostname `mbincrementaljson.local` that must
be defined in the container, for example by using the option
`--add-host` of `docker run`.

It follows the same approach as for full export dump.