Update mutate-cow to v5
In a couple of previous commits, I had removed usages of mutate-cow:

 * 98d8fbbe502fc04632e051268f3f935d1c408237
 * e056bea34cb0486532486939f18b46ea864bc603

Although the reasoning in those commits was that mutate-cow didn't help make
the code any simpler, I was in fact hesitant to continue using it elsewhere too
because of its reliance on `Proxy` objects, which I consider an unsafe, leaky
abstraction. (See the Autocomplete2 commit above and try to guess why
`unwrapProxy` was needed there, or what would have happened if you removed it.
Flow would certainly not have indicated anything was wrong!)

The issue with removing mutate-cow completely is that in the cases where it
works, it's so *useful*. So I spent some free time over the weekend releasing a
new version that removes its reliance on `Proxy` objects. Unfortunately, this
completely changes the API, which is now much more visible: instead of
receiving a `Proxy` which you can try to pretend doesn't exist (except when it
blows up in your face), you receive a "context" object which has explicit
`.get` and `.set` methods to update values.

Even though the API is a bit more verbose, using the library is not more
verbose on the whole, as this commit demonstrates: we can remove a lot of
type-level code, because separate "Writable" types are no longer required. The
typing situation is much more strict: mutate-cow no longer blindly trusts a
"Rw" type that you pass in (which could have been anything, really), and no
longer hides the fact that you're not interacting with a real object. Which
makes me a lot happier and more inclined to continue using the library in MBS!