Set -w (page width) on msgcat and msguniq
There is a persistent issue where running `./po/update_pot.sh` on different
machines will re-wrap the strings to different page widths.

This hardcodes `-w` on both msgcat and msguniq to 79, which seemed to produce
the least number of changes when I re-ran `./po/update_pot.sh`.

However, I'm not sure if this actually resolves the issue, because there may be
other tools and variables affecting things.