MBS-13378: Add preference for note emails on abstain votes
This gets rid of the weird hardcoded list of user IDs who do not get
emails for notes left on edits they abstained on.
Instead, it adds an actual preference, so that editors can choose
whether they want to use abstain to wait and see what else the editors
have to say (send emails), or to get rid of the edit
on their unvoted list and not hear about it anymore (send no emails).
If the editor hasn't yet picked an option here, the default will
be based on their "email_on_vote" preference.