MBS-13388: Handle vk.gy links
This is a visual kei database.

They have the following kinds of links I've found:
artists is for groups, so artist
musicians is for people, so artist
labels is for, well, labels
releases is for... yes, releases
lives is for events
livehouses (what?) is for places
songs is for... something in between work and recording,
since lives are the same "song" but covers are a new one.
As such, I left it unlinkable for now.