MBS-13393: Reduce the relationship data stored in sessionStorage
If an entity has many relationships and/or external links, it's possible that
saving all of them to `sessionStorage` will hit a storage quota limit, which
then throws an exception and prevents any relationship changes from being
submitted at all.

Remember that React state is saved to `sessionStorage` right before the form
submits in order to make it easier to restore if the form reloads.

In order to mitigate the issue, I've made the following changes:

 * For the relationship editor, we now only save relationships that
   actually changed (including additions) to `sessionStorage`. Previously,
   we were saving all existing relationships, even ones untouched.

 * For the external links editor, we don't appear to track which specific
   links were changed directly. However, I noticed we weren't using
   `compactEntityJson` before saving the `submittedLinks`, so now we are.

It's unlikely that a large number of external links alone would have triggered
the "quota exceeded" error. I believe that a large number of non-URL
relationships was a necessary precondition, since the relationship editor state
is much more complex and takes up more space. So the first point on its own
likely resolves the issue, but using `compactEntityJson` in both places
can't hurt.