Fix quick links key warning (#3049)
This was giving a warning that the <td> should have a key.

I looked at the code and it seemed stupid to have the <td>
in the QuickLinks return to begin with, so I changed it to return
just the links. That solves the warning, but it probably shouldn't
(possibly a React bug?) so to be actually safe, this also creates
a React fragment to return instead of the array.