Support CODE defaults in saves_preferences
If `$default` is a subroutine, it has to be called to get the default value.

But we can't call it on the existing preferences object, because it may depend
on the value of a preference which was modified (for example,
`email_on_abstain` defaults to the value of `email_on_vote`). So we first
construct the new preferences object to invoke `$default` on.