Implement focusManagement.js using the tabbable library
The `tabbable` library seems to be more robust, and making use of it means less
code for us to maintain. It's also a dependency of `floating-ui`, the successor
to `popperjs`, which means we can avoid duplicating code that achieves the same
purpose (if we upgrade to `floating-ui`).

The only downside is that `tabbable` is likely "much" slower, since it uses a
more reliable method of determining whether an element is visible (causing
layout reflow). Its API also returns every tabbable element inside the
container as an array. The previous code just tried to find the next tabbable
element using the current active element as a cursor: much less work!

But an advantage of `tabbable` is that it respects `tabIndex` ordering, unlike
the previous code. (This is likely why it needs to build the entire list of
elements as an array: in order to sort them by `tabIndex`, which is unrelated
to the document order.) And as a final note, any performance impact is unlikely
to be noticed by the user.