Upgrade from popperjs to floating-ui
Our version of popperjs is obsolete and (I believe) no longer maintained. It
was renamed to floating-ui in a major version release a long time ago, and
unfortunately came with a host of API changes. And unfortunately for us, it's
now written in TypeScript instead of Flow, and does not provide any Flow types.

I had to convert a subset of their TS types by hand in
flow-typed/npm/@floating-ui/react_v0.26.x.js.

One upside is that the new version has a much larger scope and allows us to
remove a lot of code; the size of this diff is misleading, since most of the
additions are from the ported Flow types rather than actual executable code.

Here are a few notable instances where I was able to delete code:

 * Our `useReturnFocus` hook was no longer needed, as this functionality
   is now built-in.

 * Likewise for `useOutsideClickEffect`, although it's still needed by
   our `Autocomplete2` component.

 * Much of the code in Dialog.js and focusManagement.js was redundant.
   I removed the former entirely, as it was no longer necessary to share
   any code between the `ButtonPopover` and `Modal` components. Relatedly,
   I removed the `Popover` component because it wasn't used outside of
   `ButtonPopver`.

Things should mostly look and work the same, though there is some visual
difference in how our modals are displayed: they fill up the whole screen
vertically now (see the "Add a new $entity" dialogs for reference, which is
currently the only use of Modals). The popover arrows are also now rendered as
SVGs instead of with CSS.