MBS-13273 (II): Avoid duplicating feat. artists
Fix a bug introduced by 2d82124e where tracklist artist
credits could be duplicated when changing release artists.

When changing release credits to "A & B" while a track was
already credited to "A feat. B", the track credits would be
changed to "A & B feat. B". This change sets track credits
to the new release credits (i.e. "A & B") if they already
contain the same artists as the new release credits.