Fix warning in Chrome submitting the ExternalLinkAttributeDialog
Fixes "Form submission canceled because the form is not connected".

This seems to be triggered due to the presence of the `onClick` handler on the
submit button. This is redundant, because we already have a submit handler.

I've removed the click handler, and changed the submit handler to call
`handleConfirm` instead. Also, the submit handler now unconditionally calls
both `stopPropagation` and `preventDefault` on the event: the former because we
don't want the submit event to propagate to any parent form, and the latter
because the form isn't real and shouldn't actually be submitted by the browser.