Don't store incremental JSON dumps in a local volume (#3142)
Currently, when incremental JSON dumps are generated, they are (1) copied to a
local "FTP directory" (which, confusingly, is not our actual FTP server; it's
just a local volume), and (2) rsynced to a (possibly) remote volume on the same
host as the metabrainz.org container.

There is not really any reason to do (1) if (2) is succesful, because we
presumably backup (2) already with borgmatic; there is no need to store these
in three places.

This is more relevant now, because the incremental JSON dumps are being moved
to hendrix, which is also where the metabrainz.org container is hosted. So
without these changes, we'd be storing them twice on the same host!

This updates admin/RunIncrementalJSONDump to NOT copy any dumps from the
temporary dump directory to any local volume, only rsync the dumps directly
from the temporary directory. This means we can no longer use
rsync-fullexport-files, because that runs rsync with the `--delete` flag, and
we do NOT want to delete any "remote" (but really local, now) incremental JSON
dumps that do not exist in our temporary directory (read: all of them).

I've added separate `RSYNC_INCREMENTAL_JSON_DUMPS_*` environment variables to
control the rsync invocation, which I think is clearer than trying to use
scripts designed for fullexports anyway (which are intended to only store the
few most recent dumps).