Clear the cache after each test in `t::Context`
When using `cache_aware_c` in the tests, we currently don't clear the Redis
test database between test cases (similar to how we rollback `$c->sql`).

We also don't appear to be initializing the Redis store to use the
`test_database` number.

This commit adds a separate configuration option for the minimum Redis test
database number to DBDefs.pm, so that we can clear the correct database number
for both the cache and the store after each test case.