Remove `cache_aware_c` and mock caches
Initialize the Redis cache/store in the normal `c` instead of requiring access
to a separate `cache_aware_c` in the tests. Remove the mock caches, and just
use the Redis caches, which is a better test of real code; in fact it revealed
at least one bug where we don't invalidate releases from the cache when their
ASINs are updated (fixed in the prior commit).

I didn't see any point to the `EntityCache` and `GIDEntityCache` tests having
to check how many times certain mock methods were called. It had no bearing on
what ended up being stored in the cache, which was all that mattered.