MBS-13432: Also check if non-required note is valid
Just because a note is not required, it doesn't make a useless note
(with just space, punctuation or a single ASCII character)
any less useless, so we should be blocking these
all the time.

This also fixes an issue where an empty note in the release editor
would show both the missing note and invalid note errors.