MBS-13432: Also check if subsequent edit notes are valid
If we don't check for valid edit notes in Form::Vote,
anyone can enter useless notes as long as they do it
while commenting on an edit, rather than creating it.

The fact the user is commenting doesn't seem to make a useless note
(with just space, punctuation or a single ASCII character)
any less useless.

This patch drops the invalid note, but still submits
the vote, if any. This is the same as we already do for
space-only notes that get trimmed to nothing,
so it should at least be consistent.
Since an invalid edit note is either an honest mistake
or vandalism, it seems fine not to notify the editor
about the fact that it was dropped, which would be harder
given that all the voting happens in /enter_votes
rather than the actual edit page or edit list.