MBS-13438: Restrict and validate Qobuz URLs
We supported Qobuz URLs inasmuch as we showed them
on the sidebar with their favicon, but we were not
autoselecting them, since we couldn't until recently allow
more than one option, and as such we were not
validating them, since that was not possible without the selection.

Now we can offer the following options for selection:
paid streaming (always available) or
both paid streaming and download purchase
(available in some markets, but not in LATAM for example).

We can also validate per entity: artists get /interpreter,
labels get, well, /label, and releases get /album.