MBS-13413: Allow using "Add/Remove note" with keyboard (#3133)
Since this is an <a> without href, it was being skipped
by the tabindex, but we want it to be tabbable. There seems
to be no reason why this should not be <button> instead,
which makes it tabbable by default. It also means it
responds to Enter and Space keypresses, which is nice.