MBS-13465: Prefer non-raw images when displaying front release group cover art
If a release group does not have cover art explicitly set, then we select the
first front image from the earlier release in the RG, including "Raw/Unedited"
images.

Before sorting by date, sort by the presence of "Raw/Unedited" in the cover art
types, preferring front covers that do not have this type.

Note: A similar PR is needed for the artwork-redirect API (though it doesn't
matter which is deployed first).