Drop submitting GINC/GSIN/SSCC barcode through API
GINC/GSIN (17 digits) and SSCC (18 digits) shall be used for shipment
according to the GS1 standard (subsuming EAN and UPC), not for release.

See official specifications at https://ref.gs1.org/standards/genspecs/

None of the barcodes in the database or even in tests is actually
valid with respect to GS1 specifications for GINC/GSIN/SSCC barcodes.

There are currently 101 barcodes of length 17 or 18 in the database.
Very few of these have a valid check digit, most likely just by chance.
Most of these are actually barcodes extended with five-digit add-on,
that requires a separate following patch for handling barcode add-ons.

For now this patch blocks 17/18-digit barcode submission through API.
It aligns with the release editor which never recognized these either.