Nuance the JS client-side validation of barcode
The barcode validation written in JavaScript and ran by the web browser
is tested only for the check digit of GTIN (EAN/UPC) barcode, and
is informative only (rather than mandatory).

Note that the check digit is supposed to increase the security of
reading a barcode, not to be made up when its integrity is in doubt.