Amend 9649c62a: Drop supporting GTIN-14 in tests
A GTIN-14 barcode is a 14-digit GTIN that doesn’t start with 0,
despite the misconception that any 14-digit GTIN would be a GTIN-14
as initially assumed when implementing MBS-13346 in commit 9649c62a.

GTIN-14 is meant for shipment and not for retail point of sale.

See official specifications at https://ref.gs1.org/standards/genspecs/

The database contains currently above 18,500 14-digit barcodes:

- Mainly ~18,000 of these are starting with a 0, and among those:
  - Only ~20 of 0-padded 14-digit barcodes are invalid GTIN.
- Only ~500 are left not starting with a 0, and among those:
  - Mainly ~350 are valid GTIN-13 with a 5-digit add-on,
  - Then ~110 are invalid GTIN,
  - Only ~60 are valid GTIN-14, most likely just by chance.

It isn’t worth validating the first digit specifically since most of
14-digit barcodes that don↓t start with a 0 are invalid GTIN anyway, and
since it seems likely that these follow another format, unknown for now.

It isn’t worth testing their validation either, for the same reasons,
and also because it would imply that we intentionally support GTIN-14.