Add context to [removed] strings
This is needed to be able to use the right gender and number
in translations. There's a few uses where we don't know
what is going to be used for (different entities for example
have different genders), so this at least specifies this is generic usage
so translators can choose to use a generic string here if available.