Add context to [unknown] strings (#3161)
This is needed to use the right gender in translations.
For date ones we can be specific. For the entity ones,
all we can say is that it's for a generic entity for now
(we could later decide to separate these into strings
for every entity type based on context though).