Remove reference to CAA in the footer
It was unclear whether we should mention the Event Art Archive in the footer,
too.  Since we already credit the CAA elsewhere, including prominently on the
cover art listing pages, I removed the existing CAA reference rather than make
the text in the footer (which I imagine very few people read) more bloated than
it needs to be.