MBS-13472: Also update rel-editor entity name with guess case
We update the entity name the relationship editor sees with
useEntityNameFromField so that when editors rename the entity
the current name in the form is shown on relationship previews
and the like, but this depends on an event listener
that was not getting triggered by guess case.

This just makes it so that the guess case button also triggers
an input event, so that useEntityNameFromField
will be able to use the new name.