MBS-13474: Allow TikTok videos on recordings
We were turning these into profile links, but actually
there's a fair amount of "TikTok music" being created
that originates in and is mostly streamed at
TikTok itself.
As such, it makes sense to allow linking recordings to
/video TikTok links (/music links, which are used for
videos that use licensed music, do not seem to belong
in MusicBrainz).