MBS-13481: Invalidate release on label changes
Prior to MBS-13449, cache additions were performed immediately, not after the
transaction committed. This now means that if we load additional data onto an
entity, that data will also be cached with the entity, unlike before.

This appears to be what happened with MBS-13481, as the `EditReleaseLabel`
class loads release labels onto the linked release. That wouldn't have been an
issue if we later invalidated the cached release when the labels changed, but
we didn't, because no code existed to do so.

This commit updates `Data::ReleaseLabel` to always invalidate their linked
releases whenever an insert, update, or delete occurs.

This is not the only change that's needed: there is still an issue in that
calling `$c->model('ReleaseLabel')->load` will unconditionally load fresh
release labels from the database whenever it's called, and append them onto the
existing `labels` array. Thus, if a release is loaded from the cache with
labels attached, they will be duplicated once `load` is called.

While we could just discard and overwrite the existing labels array, that would
defeat the purpose of having them cached, which is to avoid accessing the
database. To address this, I've added a `has_labels` predicate which is used to
skip releases that already have labels loaded.