Don't delay cache additions until after the transaction ends
This partially reverts 82c79f531edad396c20a2fceb45a37de9035e055 and
42be4bcf01f198530789772629093d1b69b853c5.

After searching for issues similar to MBS-13481 (where we load additional data
onto an entity which is then entered into the cache, but where changes to that
additional data do not invalidate the entity's cache entry), I believe that
there are simply far too many uncovered cases which we'd have to handle.

While storing more data in the cache and having intelligent invalidation logic
for all the linked data sounds desirable, it's too many changes to manage for
now, so we need to restore the previous behavior.

I realized (later than I should have) that we don't need to delay cache
additions until after the database transaction ends. We may have had to in a
previous implementation I had, but no longer.

The `post_txn_callbacks` ran regardless of whether or not the commit succeeded,
so an entity would always be cached in the state it was initially loaded in.
The only difference now is when that occurs. There's no type of race condition
that isn't already handled:

 * If the entity is deleted from the cache right after it's added, then there
   is no issue, just like before.

 * If the entity is added to the cache right after it's deleted, then there is
   still no issue, because our existing code to delete the added entry if there
   exists a `recently_invalidated` key for the same still applies.

The existing `*EntityCache` tests pass, and I've added a couple additional
tests to ensure that neither release events nor release labels are entered into
the cache with a release.