MBS-13487: Remind of guidelines in pseudo-release CAA tab
Cover art should generally not be added to pseudo-releases,
but there's nothing on the release page that lets the user
know about that; they'd need to go read the specific guidelines.
This adds a sentence to the cover art tab of these releases
letting the user know they shouldn't generally upload art.
This is shown only when no art exists yet for the release,
which seems the most important time and also the only one
where the text is likely to be read anyway.