MBS-13476: Show release label disambiguations on hover
We hide disambiguations for labels on the release sidebar
(I assume because of the lack of space) but this can be useful
to quickly check if the right label is selected, so we should at least
show it on hover.

This adds a showDisambiguationOnHover option to EntityLink,
which will either show the disambiguation as the whole hover
if there's nothing else there or append it in brackets if there is.