MBS-13403: Drop unused priority column from link_type
This was added in 2007 and AFAICT never used for anything.
All our relationship types had the default 0 value except for one,
which had 1 (I recently changed that to 0 as well)...
but that changed nothing since we do not use the value in any way.

I cannot drop this from the Dict in AddLinkType / EditLinkType
since it fails to load old edits without it, so I left it in
but added a comment explaining we no longer enter it.