MBS-13492: Move beginner status to a privilege flag
This stops us having to calculate beginnership every time, which
is slow and cumbersome, and implemented in at least three
different places in slightly different ways.

Every new editor added via /register will get the BEGINNER_FLAG
privilege by default, which will only be removed
when accepting edits once the number of edits + time passed
requirements are fulfilled.

We will need to run a one-off script to set the flag
on any existing beginners, added as 20240221-mbs-13492.sql.

The flag cannot be set/unset via Edit user, since even admins
should have no reason to override the default mechanism for this.

One small benefit of this is we can better control who is
a beginner for testing purposes, so we no longer need to add
dummy edits just to allow a test editor to vote.

We calculate this in _do_accept, rather than the edit queue itself,
because approvals skip the edit queue but are still relevant
for this check.