MBS-13500: Load foreign keys from file for incremental dumps
This modifies admin/RunIncrementalJSONDump and admin/cron/hourly-sitemaps.sh to
create a standalone test database from which to invoke
./script/dump_foreign_keys.pl on, and then pass the generated FK dump to to
their underlying scripts via `--foreign-keys-dump`.

This avoids the need to have foreign keys exist in the database, which was a
terrible hack, as these scripts operate on replicated mirrors.

In order to allow conveniently overriding the replication type in InitDb.pl
without passing a flag, I've allowed the default value to be read from an
environment variable first.