Add `t::script::ReplicationTest` module
This package will be used to simplify the implementation of tests that require
both a master and mirror database. I've already updated t/script/dbmirror2.t to
use it as a demonstration.

In later commits, it can also be used by our incremental JSON dump and sitemap
tests so that we no longer have to build replication packets by hand.