Add `--initial-sql` to InitDb.pl
Allows specifying which SQL script to initialize a clean database with. Prior
to this commit, admin/sql/InsertDefaultRows.sql was always used (and this could
not be disabled!).

One current use case of this is that we can initialize test databases with
t/sql/initial.sql instead; and in that file, we no longer have to worry about
disabling triggers (because the `initial-sql` script is run before triggers are
added) or specifying `ON CONFLICT` handlers (which were to avoid conflicts with
existing data that was in admin/sql/InsertDefaultRows.sql).

The triggers issue is particularly relevant for mirror databases, where certain
triggers do not exist. We can now initialize mirror test databases with
t/sql/initial.sql, where previously it would error on the `DISABLE TRIGGER
a_ins_instrument` line. But removing that line means we *must* run this script
before creating triggers, hence the lines I've reordered in create_test_db.sh,
and the changes I've made to the Selenium test scripts.