Add `--initial-sql` to InitDb.pl
Allows specifying which SQL script to initialize a clean database with. Prior
to this commit, admin/sql/InsertDefaultRows.sql was always used (and this could
not be disabled!).

One current use case of this is that we can initialize test databases with
t/sql/initial.sql instead; and in that file, we no longer have to worry about
disabling triggers (because the `initial-sql` script is run before triggers are
added) or specifying `ON CONFLICT` handlers (which were to avoid conflicts with
existing data that was in admin/sql/InsertDefaultRows.sql).