Simplify script/create_test_db.sh
With the `--initial-sql` flag to InitDb.pl, we only have to drop the existing
schemas if the database already exists (and otherwise create the pgtap
extension).

Thus, this commit delegates the majority of the work to InitDb.pl. One thing to
be aware of is that the scripts run by InitDb.pl depend on the configured
`REPLICATION_TYPE`. So if we want replication triggers, we have to set
`REPLICATION_TYPE=1`,